+++ client/vsdm-client-simservice-java/src/main/java/de/gematik/ti20/simsvc/client/repository/PoppTokenRepository.java
@@ -57,4 +57,8 @@ public class PoppTokenRepository {
     return patientProofTime != null
         && System.currentTimeMillis() / 1000 - patientProofTime > (3 * 30 * 24 * 60 * 60);
   }
+
+  public void clear() {
+    cache.clear();
+  }
 }