+++ server/vsdm-server-simservice/src/test/java/de/gematik/ti20/simsvc/server/service/FhirServiceTest.java
@@ -181,6 +181,6 @@ class FhirServiceTest {
     String result = fhirService.encodeResponse(resource, request, headers);
 
     assertNotNull(result);
-    assertTrue(headers.containsKey("Content-Type"));
+    assertTrue(headers.containsHeader("Content-Type"));
   }
 }